+++ Sources/SeriousSam/MNetworkOpen.h
@@ -1,0 +1,19 @@
+/* Copyright (c) 2002-2012 Croteam Ltd. All rights reserved. */
+
+#ifndef SE_INCL_GAME_MENU_NETWORKOPEN_H
+#define SE_INCL_GAME_MENU_NETWORKOPEN_H
+#ifdef PRAGMA_ONCE
+  #pragma once
+#endif
+
+#include "GameMenu.h"
+
+
+class CNetworkOpenMenu : public CGameMenu {
+public:
+	void Initialize_t(void);
+	void StartMenu(void);
+	void EndMenu(void);
+};
+
+#endif  /* include-once check. */
