+++ Sources/SeriousSam/MOptions.h
@@ -1,0 +1,17 @@
+/* Copyright (c) 2002-2012 Croteam Ltd. All rights reserved. */
+
+#ifndef SE_INCL_GAME_MENU_OPTIONS_H
+#define SE_INCL_GAME_MENU_OPTIONS_H
+#ifdef PRAGMA_ONCE
+  #pragma once
+#endif
+
+#include "GameMenu.h"
+
+
+class COptionsMenu : public CGameMenu {
+public:
+	void Initialize_t(void);
+};
+
+#endif  /* include-once check. */
