+++ Sources/SeriousSam/MVideoOptions.h
@@ -1,0 +1,18 @@
+/* Copyright (c) 2002-2012 Croteam Ltd. All rights reserved. */
+
+#ifndef SE_INCL_GAME_MENU_VIDEOOPTIONS_H
+#define SE_INCL_GAME_MENU_VIDEOOPTIONS_H
+#ifdef PRAGMA_ONCE
+  #pragma once
+#endif
+
+#include "GameMenu.h"
+
+
+class CVideoOptionsMenu : public CGameMenu {
+public:
+	void StartMenu(void);
+	void Initialize_t(void);
+};
+
+#endif  /* include-once check. */
