+++ Sources/SeriousSam/Menu.cpp
@@ -12,6 +12,7 @@
 #include "MenuPrinting.h"
 #include "LevelInfo.h"
 #include "VarList.h"
+#include "FileInfo.h"
 
 #include "MGArrow.h"
 #include "MGButton.h"
@@ -27,6 +28,31 @@
 #include "MGTitle.h"
 #include "MGTrigger.h"
 #include "MGVarButton.h"
+
+#include "MAudioOptions.h"
+#include "MConfirm.h"
+#include "MControls.h"
+#include "MCustomizeAxis.h"
+#include "MCustomizeKeyboard.h"
+#include "MCredits.h"
+#include "MDisabled.h"
+#include "MHighScore.h"
+#include "MInGame.h"
+#include "MLevels.h"
+#include "MMain.h"
+#include "MNetwork.h"
+#include "MNetworkJoin.h"
+#include "MNetworkOpen.h"
+#include "MNetworkStart.h"
+#include "MOptions.h"
+#include "MRenderingOptions.h"
+#include "MServers.h"
+#include "MSinglePlayer.h"
+#include "MSinglePlayerNew.h"
+#include "MSplitScreen.h"
+#include "MSplitStart.h"
+#include "MVar.h"
+#include "MVideoOptions.h"
 
 // macros for translating radio button text arrays
 #define RADIOTRANS(str) ("ETRS" str)
