+++ Sources/SeriousSam/SeriousSam.vcxproj
@@ -224,10 +224,19 @@ copy Release\$(TargetName).map $(SolutionDir)..\Bin\ &gt;nul </Command>
     <ClInclude Include="CDCheck.h" />
     <ClInclude Include="CmdLine.h" />
     <ClInclude Include="Credits.h" />
+    <ClInclude Include="FileInfo.h" />
+    <ClInclude Include="GameMenu.h" />
     <ClInclude Include="GLSettings.h" />
     <ClInclude Include="LCDDrawing.h" />
     <ClInclude Include="LevelInfo.h" />
     <ClInclude Include="MainWindow.h" />
+    <ClInclude Include="MAudioOptions.h" />
+    <ClInclude Include="MConfirm.h" />
+    <ClInclude Include="MControls.h" />
+    <ClInclude Include="MCredits.h" />
+    <ClInclude Include="MCustomizeAxis.h" />
+    <ClInclude Include="MCustomizeKeyboard.h" />
+    <ClInclude Include="MDisabled.h" />
     <ClInclude Include="Menu.h" />
     <ClInclude Include="MenuGadget.h" />
     <ClInclude Include="MenuPrinting.h" />
@@ -245,6 +254,26 @@ copy Release\$(TargetName).map $(SolutionDir)..\Bin\ &gt;nul </Command>
     <ClInclude Include="MGTitle.h" />
     <ClInclude Include="MGTrigger.h" />
     <ClInclude Include="MGVarButton.h" />
+    <ClInclude Include="MHighScore.h" />
+    <ClInclude Include="MInGame.h" />
+    <ClInclude Include="MLevels.h" />
+    <ClInclude Include="MLoadSave.h" />
+    <ClInclude Include="MMain.h" />
+    <ClInclude Include="MNetwork.h" />
+    <ClInclude Include="MNetworkJoin.h" />
+    <ClInclude Include="MNetworkOpen.h" />
+    <ClInclude Include="MNetworkStart.h" />
+    <ClInclude Include="MOptions.h" />
+    <ClInclude Include="MPlayerProfile.h" />
+    <ClInclude Include="MRenderingOptions.h" />
+    <ClInclude Include="MSelectPlayers.h" />
+    <ClInclude Include="MServers.h" />
+    <ClInclude Include="MSinglePlayer.h" />
+    <ClInclude Include="MSinglePlayerNew.h" />
+    <ClInclude Include="MSplitScreen.h" />
+    <ClInclude Include="MSplitStart.h" />
+    <ClInclude Include="MVar.h" />
+    <ClInclude Include="MVideoOptions.h" />
     <ClInclude Include="resource.h" />
     <ClInclude Include="SeriousSam.h" />
     <ClInclude Include="SplashScreen.h" />
